refactor(x/data): update data module to use orm
