+++ x/data/server/query_server.go
@@ -7,295 +7,12 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
-	"github.com/cosmos/cosmos-sdk/store/prefix"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/query"
 
-	api "github.com/regen-network/regen-ledger/api/regen/data/v1"
-	"github.com/regen-network/regen-ledger/types"
-	"github.com/regen-network/regen-ledger/types/ormutil"
 	"github.com/regen-network/regen-ledger/x/data"
 )
 
 var _ data.QueryServer = serverImpl{}
-
-// ByIRI queries data based on its ContentHash.
-func (s serverImpl) ByIRI(ctx context.Context, request *data.QueryByIRIRequest) (*data.QueryByIRIResponse, error) {
-	id, err := s.getID(ctx, request.Iri)
-	if err != nil {
-		return nil, err
-	}
-
-	store := sdk.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	entry, err := s.getEntry(store, id)
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryByIRIResponse{
-		Entry: entry,
-	}, nil
-}
-
-// ByHash queries data based on ContentHash.
-func (s serverImpl) ByHash(ctx context.Context, request *data.QueryByHashRequest) (*data.QueryByHashResponse, error) {
-	if request.ContentHash == nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("content hash cannot be empty")
-	}
-
-	iri, err := request.ContentHash.ToIRI()
-	if err != nil {
-		return nil, err
-	}
-
-	id, err := s.getID(ctx, iri)
-	if err != nil {
-		return nil, err
-	}
-
-	store := sdk.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	entry, err := s.getEntry(store, id)
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryByHashResponse{
-		Entry: entry,
-	}, nil
-}
-
-// ByAttestor queries data based on attestor.
-func (s serverImpl) ByAttestor(ctx context.Context, request *data.QueryByAttestorRequest) (*data.QueryByAttestorResponse, error) {
-	addr, err := sdk.AccAddressFromBech32(request.Attestor)
-	if err != nil {
-		return nil, err
-	}
-
-	store := types.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	attestorIDStore := prefix.NewStore(store, AttestorIDIndexPrefix(addr))
-
-	var entries []*data.ContentEntry
-	pageRes, err := query.Paginate(attestorIDStore, request.Pagination, func(key []byte, value []byte) error {
-		entry, err := s.getEntry(store, key)
-		if err != nil {
-			return err
-		}
-
-		entries = append(entries, entry)
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryByAttestorResponse{
-		Entries:    entries,
-		Pagination: pageRes,
-	}, nil
-}
-
-// IRIByHash queries IRI based on ContentHash.
-func (s serverImpl) IRIByHash(ctx context.Context, request *data.QueryIRIByHashRequest) (*data.QueryIRIByHashResponse, error) {
-	if request.ContentHash == nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("content hash cannot be empty")
-	}
-
-	iri, err := request.ContentHash.ToIRI()
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryIRIByHashResponse{
-		Iri: iri,
-	}, nil
-}
-
-// HashByIRI queries ContentHash based on IRI.
-func (s serverImpl) HashByIRI(ctx context.Context, request *data.QueryHashByIRIRequest) (*data.QueryHashByIRIResponse, error) {
-	hash, err := data.ParseIRI(request.Iri)
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryHashByIRIResponse{
-		ContentHash: hash,
-	}, nil
-}
-
-// AttestorsByIRI queries attestors based on IRI.
-func (s serverImpl) AttestorsByIRI(ctx context.Context, request *data.QueryAttestorsByIRIRequest) (*data.QueryAttestorsByIRIResponse, error) {
-	id, err := s.getID(ctx, request.Iri)
-	if err != nil {
-		return nil, err
-	}
-
-	store := types.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	attestorIDStore := prefix.NewStore(store, IDAttestorIndexPrefix(id))
-
-	var attestors []string
-	pageRes, err := query.Paginate(attestorIDStore, request.Pagination, func(key []byte, value []byte) error {
-		attestors = append(attestors, sdk.AccAddress(key).String())
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryAttestorsByIRIResponse{
-		Attestors:  attestors,
-		Pagination: pageRes,
-	}, nil
-}
-
-// AttestorsByHash queries attestors based on ContentHash.
-func (s serverImpl) AttestorsByHash(ctx context.Context, request *data.QueryAttestorsByHashRequest) (*data.QueryAttestorsByHashResponse, error) {
-	if request.ContentHash == nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("content hash cannot be empty")
-	}
-
-	iri, err := request.ContentHash.ToIRI()
-	if err != nil {
-		return nil, err
-	}
-
-	id, err := s.getID(ctx, iri)
-	if err != nil {
-		return nil, err
-	}
-
-	store := types.UnwrapSDKContext(ctx).KVStore(s.storeKey)
-	attestorIDStore := prefix.NewStore(store, IDAttestorIndexPrefix(id))
-
-	var attestors []string
-	pageRes, err := query.Paginate(attestorIDStore, request.Pagination, func(key []byte, value []byte) error {
-		attestors = append(attestors, sdk.AccAddress(key).String())
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return &data.QueryAttestorsByHashResponse{
-		Attestors:  attestors,
-		Pagination: pageRes,
-	}, nil
-}
-
-// ResolversByIRI queries resolvers based on IRI.
-func (s serverImpl) ResolversByIRI(ctx context.Context, request *data.QueryResolversByIRIRequest) (*data.QueryResolversByIRIResponse, error) {
-	id, err := s.getID(ctx, request.Iri)
-	if err != nil {
-		return nil, err
-	}
-
-	if request.Pagination == nil {
-		request.Pagination = &query.PageRequest{}
-	}
-
-	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
-	if err != nil {
-		return nil, err
-	}
-	it, err := s.stateStore.DataResolverTable().
-		List(ctx, api.DataResolverPrimaryKey{}.WithId(id),
-			ormlist.Paginate(pg))
-	if err != nil {
-		return nil, err
-	}
-
-	res := &data.QueryResolversByIRIResponse{}
-	for it.Next() {
-		item, err := it.Value()
-		if err != nil {
-			return nil, err
-		}
-
-		resolverInfo, err := s.stateStore.ResolverInfoTable().Get(ctx, item.ResolverId)
-		if err != nil {
-			return nil, err
-		}
-
-		res.ResolverUrls = append(res.ResolverUrls, resolverInfo.Url)
-	}
-
-	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// ResolversByHash queries resolvers based on ContentHash.
-func (s serverImpl) ResolversByHash(ctx context.Context, request *data.QueryResolversByHashRequest) (*data.QueryResolversByHashResponse, error) {
-	if request.ContentHash == nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrap("content hash cannot be empty")
-	}
-
-	iri, err := request.ContentHash.ToIRI()
-	if err != nil {
-		return nil, err
-	}
-
-	id, err := s.getID(ctx, iri)
-	if err != nil {
-		return nil, err
-	}
-
-	if request.Pagination == nil {
-		request.Pagination = &query.PageRequest{}
-	}
-
-	pg, err := ormutil.GogoPageReqToPulsarPageReq(request.Pagination)
-	if err != nil {
-		return nil, err
-	}
-	it, err := s.stateStore.DataResolverTable().
-		List(ctx, api.DataResolverPrimaryKey{}.WithId(id),
-			ormlist.Paginate(pg))
-	if err != nil {
-		return nil, err
-	}
-
-	res := &data.QueryResolversByHashResponse{}
-	for it.Next() {
-		item, err := it.Value()
-		if err != nil {
-			return nil, err
-		}
-
-		resolverInfo, err := s.stateStore.ResolverInfoTable().Get(ctx, item.ResolverId)
-		if err != nil {
-			return nil, err
-		}
-
-		res.ResolverUrls = append(res.ResolverUrls, resolverInfo.Url)
-	}
-
-	res.Pagination, err = ormutil.PulsarPageResToGogoPageRes(it.PageResponse())
-	if err != nil {
-		return nil, err
-	}
-
-	return res, nil
-}
-
-// ResolverInfo queries information about a resolved based on url.
-func (s serverImpl) ResolverInfo(ctx context.Context, request *data.QueryResolverInfoRequest) (*data.QueryResolverInfoResponse, error) {
-	res, err := s.stateStore.ResolverInfoTable().GetByUrl(ctx, request.Url)
-	if err != nil {
-		return nil, err
-	}
-
-	acct := sdk.AccAddress(res.Manager)
-
-	return &data.QueryResolverInfoResponse{
-		Id:      res.Id,
-		Manager: acct.String(),
-	}, nil
-}
 
 func (s serverImpl) getID(ctx context.Context, iri string) ([]byte, error) {
 	store := sdk.UnwrapSDKContext(ctx).KVStore(s.storeKey)
@@ -307,16 +24,16 @@ func (s serverImpl) getID(ctx context.Context, iri string) ([]byte, error) {
 	return id, nil
 }
 
-func (s serverImpl) getEntry(store sdk.KVStore, id []byte) (*data.ContentEntry, error) {
-	bz := store.Get(AnchorTimestampKey(id))
-	if len(bz) == 0 {
-		return nil, status.Error(codes.NotFound, "entry not found")
-	}
+func (s serverImpl) getEntry(ctx context.Context, store sdk.KVStore, id []byte) (*data.ContentEntry, error) {
 
-	var timestamp gogotypes.Timestamp
-	err := timestamp.Unmarshal(bz)
+	dataAnchor, err := s.stateStore.DataAnchorTable().Get(ctx, id)
 	if err != nil {
 		return nil, err
+	}
+
+	timestamp := &gogotypes.Timestamp{
+		Seconds: dataAnchor.Timestamp.Seconds,
+		Nanos:   dataAnchor.Timestamp.Nanos,
 	}
 
 	iri := string(s.iriIDTable.GetValue(store, id))
@@ -326,7 +43,7 @@ func (s serverImpl) getEntry(store sdk.KVStore, id []byte) (*data.ContentEntry, 
 	}
 
 	entry := &data.ContentEntry{
-		Timestamp: &timestamp,
+		Timestamp: timestamp,
 		Iri:       iri,
 		Hash:      contentHash,
 	}
